requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
python-decouple>=3.8
dj-database-url>=2.1
gunicorn>=21.2
whitenoise>=6.5
